删除微博

+++ app/Http/Controllers/StatusesController.php
@@ -28,6 +28,9 @@ class StatusesController extends Controller
 
     public function destroy(Status $status)
     {
+        $this->authorize('destroy', $status);
         $status->delete();
+        session()->flash('success', '删除成功！');
+        return redirect()->back();
     }
 }
